requirements.txt
stable-ts
google-generativeai
python-dotenv
pandas
pykakasi
PyDrive2
openpyxl
